+++ packages/finschia/src/modules/wasm/aminomessages.spec.ts
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import { fromBase64 } from "@cosmjs/encoding";
 import { AminoTypes } from "@cosmjs/stargate";
-import { MsgStoreCode } from "lbmjs-types/cosmwasm/wasm/v1/tx";
-import { AccessType } from "lbmjs-types/cosmwasm/wasm/v1/types";
+import { MsgStoreCode } from "cosmjs-types/cosmwasm/wasm/v1/tx";
+import { AccessType } from "cosmjs-types/cosmwasm/wasm/v1/types";
 
 import { AminoMsgStoreCode, createWasmAminoConverters } from "./aminomessages";