+++ packages/finschia/src/modules/wasmplus/aminomessages.ts
@@ -1,7 +1,8 @@
+/* eslint-disable @typescript-eslint/naming-convention */
 import { AminoMsg } from "@cosmjs/amino";
 import { fromBase64, fromUtf8, toBase64, toUtf8 } from "@cosmjs/encoding";
 import { AminoConverters } from "@cosmjs/stargate";
-import { Coin } from "lbmjs-types/cosmos/base/v1beta1/coin";
+import { Coin } from "cosmjs-types/cosmos/base/v1beta1/coin";
 import { MsgStoreCodeAndInstantiateContract } from "lbmjs-types/lbm/wasm/v1/tx";
 
 import { AccessConfig, accessTypeFromString, accessTypeToString } from "../wasm/aminomessages";